moved enhanced kyc and document selector to lib
