+++ sample/src/main/java/com/smileidentity/sample/Utils.kt
@@ -6,7 +6,7 @@ import androidx.annotation.StringRes
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.SnackbarResult
 import com.smileidentity.models.JobType
-import com.smileidentity.sample.compose.components.SearchableInputFieldItem
+import com.smileidentity.compose.components.SearchableInputFieldItem
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 
@@ -66,67 +66,10 @@ val JobType.label: Int
     get() = when (this) {
         JobType.SmartSelfieEnrollment -> com.smileidentity.R.string.si_smart_selfie_enrollment_product_name // ktlint-disable max-line-length
         JobType.SmartSelfieAuthentication -> com.smileidentity.R.string.si_smart_selfie_authentication_product_name // ktlint-disable max-line-length
-        JobType.EnhancedKyc -> R.string.enhanced_kyc_product_name
+        JobType.EnhancedKyc -> com.smileidentity.R.string.si_enhanced_kyc_product_name
         JobType.BiometricKyc -> com.smileidentity.R.string.si_biometric_kyc_product_name
         JobType.DocumentVerification -> com.smileidentity.R.string.si_doc_v_product_name
         JobType.BVN -> com.smileidentity.R.string.si_bvn_product_name
         JobType.EnhancedDocumentVerification -> com.smileidentity.R.string.si_enhanced_docv_product_name // ktlint-disable max-line-length
         JobType.Unknown -> -1
     }
-
-val countryDetails = mapOf(
-    "AO" to SearchableInputFieldItem("AO", "Angola", "🇦🇴"),
-    "BF" to SearchableInputFieldItem("BF", "Burkina Faso", "🇧🇫"),
-    "BI" to SearchableInputFieldItem("BI", "Burundi", "🇧🇮"),
-    "BJ" to SearchableInputFieldItem("BJ", "Benin", "🇧🇯"),
-    "BW" to SearchableInputFieldItem("BW", "Botswana", "🇧🇼"),
-    "CD" to SearchableInputFieldItem("CD", "Congo (DRC)", "🇨🇩"),
-    "CF" to SearchableInputFieldItem("CF", "Central African Republic", "🇨🇫"),
-    "CG" to SearchableInputFieldItem("CG", "Congo", "🇨🇬"),
-    "CI" to SearchableInputFieldItem("CI", "Côte d'Ivoire", "🇨🇮"),
-    "CM" to SearchableInputFieldItem("CM", "Cameroon", "🇨🇲"),
-    "CV" to SearchableInputFieldItem("CV", "Cabo Verde", "🇨🇻"),
-    "DJ" to SearchableInputFieldItem("DJ", "Djibouti", "🇩🇯"),
-    "EG" to SearchableInputFieldItem("EG", "Egypt", "🇪🇬"),
-    "EH" to SearchableInputFieldItem("EH", "Western Sahara", "🇪🇭"),
-    "ER" to SearchableInputFieldItem("ER", "Eritrea", "🇪🇷"),
-    "ET" to SearchableInputFieldItem("ET", "Ethiopia", "🇪🇹"),
-    "GA" to SearchableInputFieldItem("GA", "Gabon", "🇬🇦"),
-    "GH" to SearchableInputFieldItem("GH", "Ghana", "🇬🇭"),
-    "GM" to SearchableInputFieldItem("GM", "Gambia", "🇬🇲"),
-    "GN" to SearchableInputFieldItem("GN", "Guinea", "🇬🇳"),
-    "GQ" to SearchableInputFieldItem("GQ", "Equatorial Guinea", "🇬🇶"),
-    "GW" to SearchableInputFieldItem("GW", "Guinea-Bissau", "🇬🇼"),
-    "KE" to SearchableInputFieldItem("KE", "Kenya", "🇰🇪"),
-    "KM" to SearchableInputFieldItem("KM", "Comoros", "🇰🇲"),
-    "LR" to SearchableInputFieldItem("LR", "Liberia", "🇱🇷"),
-    "LS" to SearchableInputFieldItem("LS", "Lesotho", "🇱🇸"),
-    "LY" to SearchableInputFieldItem("LY", "Libya", "🇱🇾"),
-    "MA" to SearchableInputFieldItem("MA", "Morocco", "🇲🇦"),
-    "MG" to SearchableInputFieldItem("MG", "Madagascar", "🇲🇬"),
-    "ML" to SearchableInputFieldItem("ML", "Mali", "🇲🇱"),
-    "MR" to SearchableInputFieldItem("MR", "Mauritania", "🇲🇷"),
-    "MU" to SearchableInputFieldItem("MU", "Mauritius", "🇲🇺"),
-    "MW" to SearchableInputFieldItem("MW", "Malawi", "🇲🇼"),
-    "MZ" to SearchableInputFieldItem("MZ", "Mozambique", "🇲🇿"),
-    "NA" to SearchableInputFieldItem("NA", "Namibia", "🇳🇦"),
-    "NE" to SearchableInputFieldItem("NE", "Niger", "🇳🇪"),
-    "NG" to SearchableInputFieldItem("NG", "Nigeria", "🇳🇬"),
-    "RW" to SearchableInputFieldItem("RW", "Rwanda", "🇷🇼"),
-    "SC" to SearchableInputFieldItem("SC", "Seychelles", "🇸🇨"),
-    "SD" to SearchableInputFieldItem("SD", "Sudan", "🇸🇩"),
-    "SL" to SearchableInputFieldItem("SL", "Sierra Leone", "🇸🇱"),
-    "SN" to SearchableInputFieldItem("SN", "Senegal", "🇸🇳"),
-    "SO" to SearchableInputFieldItem("SO", "Somalia", "🇸🇴"),
-    "SS" to SearchableInputFieldItem("SS", "South Sudan", "🇸🇸"),
-    "ST" to SearchableInputFieldItem("ST", "São Tomé and Príncipe", "🇸🇹"),
-    "SZ" to SearchableInputFieldItem("SZ", "Eswatini", "🇸🇿"),
-    "TD" to SearchableInputFieldItem("TD", "Chad", "🇹🇩"),
-    "TG" to SearchableInputFieldItem("TG", "Togo", "🇹🇬"),
-    "TN" to SearchableInputFieldItem("TN", "Tunisia", "🇹🇳"),
-    "TZ" to SearchableInputFieldItem("TZ", "Tanzania", "🇹🇿"),
-    "UG" to SearchableInputFieldItem("UG", "Uganda", "🇺🇬"),
-    "ZA" to SearchableInputFieldItem("ZA", "South Africa", "🇿🇦"),
-    "ZM" to SearchableInputFieldItem("ZM", "Zambia", "🇿🇲"),
-    "ZW" to SearchableInputFieldItem("ZW", "Zimbabwe", "🇿🇼"),
-)
