+++ sample/src/main/res/values/strings.xml
@@ -51,30 +51,7 @@
     <string name="user_id_dialog_title">Enter User ID</string>
     <string name="user_id_dialog_supporting_text">From a previous successful SmartSelfie™ Enrollment</string>
 
-    <!--  Enhanced KYC  -->
-    <string name="enhanced_kyc_product_name">Enhanced KYC</string>
-    <string name="enhanced_kyc_instructions">Provide ID Information</string>
-    <string name="enhanced_kyc_country_picker_label">Country</string>
-    <string name="enhanced_kyc_id_type_picker_label">ID Type</string>
-    <string name="enhanced_kyc_submit_button">Submit</string>
-    <string name="enhanced_kyc_processing_title">Processing Enhanced KYC…</string>
-    <string name="enhanced_kyc_processing_subtitle">Just a few more seconds</string>
-    <string name="enhanced_kyc_processing_success_title">Enhanced KYC Complete</string>
-    <string name="enhanced_kyc_processing_success_subtitle">You can now proceed</string>
-    <string name="enhanced_kyc_processing_error_title">Enhanced KYC Failed</string>
-    <string name="enhanced_kyc_processing_error_subtitle">This could be because of internet connectivity</string>
-    <string name="enhanced_kyc_processing_continue_button">Continue</string>
-    <string name="enhanced_kyc_processing_retry_button">Retry</string>
-    <string name="enhanced_kyc_processing_close_button">Close</string>
-
-    <!--  Biometric KYC  -->
-    <string name="biometric_kyc_instructions">Provide ID Information</string>
-    <string name="biometric_kyc_enter_id_info">Enter ID Information</string>
-
     <!--  Document Verification  -->
-    <string name="doc_v_select_country">Select Country of Issue</string>
-    <string name="doc_v_country_search_field_hint">Select Country</string>
-    <string name="doc_v_select_id_type">Select ID Type</string>
     <string name="doc_v_info_title">Get Your ID Verified</string>
     <string name="doc_v_info_subtitle">We need a valid document to verify who you are.</string>
 </resources>
